cleaned dead code
