+++ src/helpers.rs
@@ -1,0 +1,60 @@
+use serde_json::map::Map;
+use serde_json::value::Value;
+
+pub fn flatten(v: Value) -> Value {
+    match v {
+        Value::Object(map) => Value::Object(flatten_value_map(map)),
+        _ => v,
+    }
+}
+
+pub fn deflatten(v: Value) -> Value {
+    match v {
+        Value::Object(map) => Value::Object(deflatten_value_map(map)),
+        _ => v,
+    }
+}
+
+fn flatten_value_map(map: Map<String, Value>) -> Map<String, Value> {
+    let mut result = Map::new();
+    for (key, value) in map {
+        match value {
+            Value::Object(obj) => for (k, v) in obj {
+                let new_key = format!("{}.{}", key, k);
+                result.insert(new_key, flatten(v));
+            },
+            _ => {
+                result.insert(key, value);
+                ()
+            }
+        };
+    }
+    result
+}
+
+fn deflatten_value_map(map: Map<String, Value>) -> Map<String, Value> {
+    let mut temp = Map::new();
+    let mut result = Map::new();
+    for (key, value) in map {
+        let mut iterator = key.splitn(2, '.');
+        if let Some(newkey) = iterator.next() {
+            match iterator.next() {
+                None => {
+                    result.entry(newkey).or_insert(value);
+                    ()
+                }
+                Some(subkey) => {
+                    let submap = temp.entry(newkey).or_insert(Value::Object(Map::new()));
+                    if let Some(submap) = submap.as_object_mut() {
+                        submap.entry(subkey).or_insert(value);
+                    }
+                }
+            }
+        }
+    }
+
+    for (key, value) in temp {
+        result.insert(key, deflatten(value));
+    }
+    result
+}
